solution

+++ 1406B.cpp
@@ -5,13 +5,12 @@ int main(){
 	int tc;cin>>tc;
 	while(tc--){
 		int n;cin>>n;
-		int arr[n];
-		for (int i = 0; i < n; i++)
+		long long arr[n];
+		for (long long i = 0; i < n; i++)
 		{
 			cin>>arr[i];
 		}
 		sort(arr,arr+n);
-		// reverse(arr,arr+n);
 		long long ans1,ans2,ans3;
 		ans1=arr[0]*arr[1]*arr[2]*arr[3]*arr[n-1];
 		ans2=arr[0]*arr[1]*arr[n-1]*arr[n-2]*arr[n-3];
